+++ src/gui/options/OptionsController.h
@@ -53,7 +53,7 @@ public:
         // TM-mode EM field settings (EMWave2 port)
         void SetEMEnabled(bool state);
         void SetEMCellSize(int cellSize);
-        void SetEMRegionScale(int regionScale);
+        void SetEMRegionScale(float regionScale);
         void SetEMBoundaryMode(int boundaryMode);
         void SetEMSourceMode(int sourceMode);
         void SetEMFrequency(float frequency);
